WebsiteBackend/requirements.txt
annotated-types==0.6.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1

bleach==6.1.0

blis==0.7.11
breadability==0.1.20
cachelib==0.9.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.0
confection==0.1.4
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-mongodb-engine==0.6.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.1.0
docopt==0.6.2

et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0

Flask==3.0.0
Flask-Caching==2.1.0
fonttools==4.45.1
fqdn==1.5.1
gensim==4.3.2

h11==0.12.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.13.3
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.11.0
imblearn==0.0
ipykernel==6.26.0
ipython==8.18.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1

kiwisolver==1.4.5
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3

mistune==3.0.2
mongoengine==0.27.0
murmurhash==1.0.10
mysqlclient==2.2.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1

notebook_shim==0.2.3
numpy==1.26.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3

phonenumbers==8.13.26
Pillow==10.1.0
pipreqs==0.4.13
platformdirs==4.0.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycountry==23.12.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyJWT==2.3.0
pylance==0.8.17
pymongo==3.11.4
pyparsing==3.1.1
pypdfium2==4.24.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2021.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
resolver==0.2.1
resume-parser==0.8.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.13.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post11
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.2.4
srsly==2.4.8
stack-data==0.6.3
stemming==1.0.1
sumy==0.11.0
terminado==0.18.0
thinc==8.2.1
threadpoolctl==3.2.0
tika==2.6.0

tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2021.5
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
wordcloud==1.9.2
yarg==0.1.9
